+++ merlin/common/tasks.py
@@ -6,7 +6,7 @@
 #
 # LLNL-CODE-797170
 # All rights reserved.
-# This file is part of Merlin, Version: 1.12.0.
+# This file is part of Merlin, Version: 1.12.1.
 #
 # For details, see https://github.com/LLNL/merlin.
 #
@@ -49,7 +49,7 @@ from merlin.config.utils import Priority, get_priority
 from merlin.exceptions import HardFailException, InvalidChainException, RestartException, RetryException
 from merlin.router import stop_workers
 from merlin.spec.expansion import parameter_substitutions_for_cmd, parameter_substitutions_for_sample
-from merlin.study.status import read_status
+from merlin.study.status import read_status, status_conflict_handler
 from merlin.utils import dict_deep_merge
 
 
@@ -484,7 +484,7 @@ def gather_statuses(
                         # Make sure the status for this sample workspace is in a finished state (not initialized or running)
                         if status[step_name][f"{condensed_workspace}/{path}"]["status"] not in ("INITIALIZED", "RUNNING"):
                             # Add the status data to the statuses we'll write to the condensed file and remove this status file
-                            dict_deep_merge(condensed_statuses, status)
+                            dict_deep_merge(condensed_statuses, status, conflict_handler=status_conflict_handler)
                             files_to_remove.append(status_filepath)
                             files_to_remove.append(lock_filepath)  # Remove the lock files as well as the status files
                     except KeyError:
@@ -556,7 +556,7 @@ def condense_status_files(self, *args: Any, **kwargs: Any) -> ReturnCode:  # pyl
                         existing_condensed_statuses = json.load(condensed_status_file)
                     # Merging the statuses we're condensing into the already existing statuses
                     # because it's faster at scale than vice versa
-                    dict_deep_merge(existing_condensed_statuses, condensed_statuses)
+                    dict_deep_merge(existing_condensed_statuses, condensed_statuses, conflict_handler=status_conflict_handler)
                     condensed_statuses = existing_condensed_statuses
 
                 # Write the condensed statuses to the condensed status file